uclibc: remove forced gcc optimization
It was added in 6d6bd8ba78434ecb09395582b5f3e41febd4d4ee,
but it is unclear for me, why this is needed or if it is
required. I don't think we should depend on some gcc
optimization, which might change in the future.
Anyway, this breaks c6x toolchain building (ICE), so I remove it.
Testsuite run for supported architectures didn't add any
new errors.

Change-Id: Iedd203f85a124a74fc79abb93789deb4167b30d9