uclibc/c6x: Add note on alternate tree
This is a squash of:

c6x: disable it at the moment

There is a NPTL port of uCLibc for c6x, but it doesn't work
with uClibc master. May be someone with real hardware could
sync it. Unfortunately I got no respone from Aurelien and
no response on the mailing list.

reenable c6x for toolchain testing without NPTL

Change-Id: I2f44a9a516f9fcbdd40a65bc297a8f0e1e0095b5