uclibc/mkostemp: fix implementation
mkostemp(char *template, int flags) generates a unique temporary
filename from a template.  The flags parameter accepts three of
the same flags as open(2): O_APPEND, O_CLOEXEC, and O_SYNC.  The
current implementation of mkostemp(3) does not respect the flags
and in fact confuses the flags with the file mode which should
always be S_IRUSR | S_IWUSR.  This patch corrects this issue.

Change-Id: I4984f2549be2c10e9eb126eed139ca755640ef4f
Signed-off-by: Anthony G. Basile <blueness@gentoo.org>