uclibc/stdio: Fix printing 0.0
We were relying on FE_DIVBYZERO being turned off when printing
"%f", +-.0
Avoid the whole issue by looking at the sign-bit (in a rough
approximation).

Note that we do not handle gracefully:
  printf ("\n%llf\n", -0.0);
  printf ("\n%llf\n",  0.0);
nor %Lf for both when NOT cast to long double. Avoiding an FPE due to
broken numbers like these does not make sense to me.

Change-Id: I2c3b7c2b131f91cb5c900dc0e27b90149e7b2e6f
Signed-off-by: Bernhard Reutner-Fischer <rep.dot.nop@gmail.com>