uclibc/ldso: mark _dl_exit as noreturn
Otherwise gcc might not understand that oom() ended control-flow and
might emit an (untaken) reference to abort() in _dl_update_slotinfo()
on e.g. SH4 which breaks linking ld-uClibc.so.
Arguably -ffreestanding should prevent GCC from emitting this
'.global abort' but alas, it does not, which is another bug..

Also mark the function cold to further lower the incoming frequency and
branch probability.

Change-Id: I06d6234a95769d381ba6b67d7491cc24182b886e
Signed-off-by: Bernhard Reutner-Fischer <rep.dot.nop@gmail.com>