Ned: fix life-time of caps in app model
We have to make sure that we keep a copy of all ref-counted caps that we
intend to map into our application until we actually mapped them to the
new task. Otherwise the Lua GC might just collect the Lua cap and this
might contain the last copy of the ref-counted cap.

So this change adds a simple container that stores ref-counted caps and
drops them when the application is running.

Change-Id: Ia32504f0710cfd895e2104384d8ebb64a8935f21