uclibc: Replace bltzal with lapc for MIPS R6 builds.
The R6 version of MIPS does not support the bltzal instruction.  This
patch changes dl-startup.h and dl-sysdep.h to use lapc on R6 instead.
lapc is a new R6 insruction so older systems need to continue to use
bltzal in order to load register $31.

Signed-off-by: Steve Ellcey <sellcey@imgtec.com>

Also includes the fixup:

ld.so: fix mips{32,64}r6 support

Signed-off-by: Matthew Fortune <Matthew.Fortune@imgtec.com>

---

Change-Id: Ic52d60f552af3cdf2ed4e3c617897072fa772153