uclibc: pthread_atfork handlers not removed during dlclose
Invoke pthread_atfork handler cleanup when removing the associated
DSO...

If a program loads a DSO (dlopen) that sets up a pthread_atfork
handler(s), and then subsequently closes the DSO, the handler(s) are
left in place.  If fork() is subsequently called, the handlers are
invoked even though the DSO has been removed causing crashes or
unpredictable code execution.  This is because the code in
__cxa_finalize(atexit.c)to invoke the unregister_atfork() routine is
ifdef'd out with the comment that it hasn't been "looked into this
yet...".

Refs.:
 http://bugs.busybox.net/show_bug.cgi?id=8211
 http://sourceware.org/bugzilla/show_bug.cgi?id=13502

Add test-case, enable cleanup for NPTL only.

Change-Id: I153a3aebb9fae23bfb2a6604487d1371af696ae5
Signed-off-by: John Ata <john.ata@baesystems.com>
Signed-off-by: Leonid Lisovskiy <lly.dev@gmail.com>