l4sys: Add L4::Registry_iface
L4::Registry_iface provides an abstract interface for object registries
that allow to register and unregister server-side Epiface objects for
either RPC or IRQs.

L4Re::Util::Object_registry implements this interface, so now
this abstract interface might be passed to generic modules that need to
register server-side object at a server loop without generating a
dependency to the exact implementation of the server-loop and registry
itself.

Change-Id: I82e7049b9d8193b52f9696ba34d40266cb57bdc0