test: Scheduler migrating a thread to online CPUs
The test migrates one thread to all online cores. On each core the thread
executes on, it consumes time. The consumed time of the thread since the last
migration is prove, it executed on each core.
The test fails, if the thread is migrated to an online core, where it does not
consume cpu time.

To ensure it actually executed an IPC is sent to the thread allowing it
to execute.

Requires-Change-Id: I61e865428837a6f4cc343c3a4602993b7c3b86c2
Change-Id: I705fd1ae8af84e624aad1ef242872562f6de1ae2