util: add helper macro to define a thread startup wrapper
The wrapper is normally empty but can be used for architectures
where additional setup (apart from setting SP and IP) is required
to get the CPU into a state that is compatible with C calling
conventions.

Adds a custom implementation for MIPS which requires GP and t9
to be set up.

Change-Id: I35765059f38a5f8bfc83c16d46f9f95e58440672