l4sys: add ARCH specific ex_regs trigger exception test
This change adds a function to utcb.h testing a given exception if it
was caused by L4::Thread::ex_regs with the
L4_THREAD_EX_REGS_TRIGGER_EXCEPTION flag.

PPC and SPARC: The kernel does currently not implement
the code path emitting the exception and therefore the test function
returns 0.

Change-Id: Idb0dc4deb366633a933ab525f4fbef0b38ff9e89