l4sys: adds l4_fpage_memaddr, l4_fpage_ioport, l4_fpage_obj
l4_fpage_page provides the page part of the given flex page. In case
of a memory flex page this is the page frame number and not the page
address. The caller needs to shift the result depending on the fpage
type with the correct value.

The type specific l4_fpage_memaddr/ioport/obj functions provide memory
address, IO port number, or capability index, respectively. This
encapsulates the flex page internal shifting.

Change-Id: I3677405c948b0b2e0109b2f4a8259e97c9e734c4