ned: correctly propagate requested rights when registering in namespace
When registering a capability in a namespace, the rights must not
only be set in the flags for the function but also for the capability
to be registered.

Keep the default right for empty capabilities at read-only, as it has
been effectively now because no explicit right has been set on the
capability.

Change-Id: Iea48b23de4eea4928ad58dbf04c8eff853e3a23f