uclibc: pthread mgr: keep cap_slot allocated on thread exit
Before this change, the pthread manager deleted the kernel thread
object and freed the capability slot before pthread_join() or
pthread_detach() were called. This leads to a race condition, where
the capability slot of the still existing pthread could be
reallocated by another pthread. Subsequently, requesting the
capability of the old pthread returns a valid capability index
with a present kernel object.

Furthermore, consider a second case where the creator of the
pthread checks the capability index for validity followed by
an IPC call with infinite timeout. If the pthread exits in
between these two statements and the capability is invalidated,
the creator will issue an IPC on itself and block forever.

This changes issues just the deletion request for the kernel objects
on the pthread exit. It keeps the capability slot allocated until
the pthread is either joined or detached, which invokes pthread_free.
Pthread_free then frees the capability slot.

This way, both race conditions are mitigated, as the capability slot
cannot be reallocated and an IPC to a not-present kernel object
returns with an error instead of blocking forever.

Change-Id: I41945c83d2dab93b33d9f12831561696b9ff9974