l4sys: vcon: change buffer size param type to unsigned
Vcon's send, write, read, and read_with_flags functions have a size
parameter of type 'int' describing the size of the passed buffer.

Changing the type to 'unsigned' documents the specified value range
for these functions. Furthermore, it eliminates an implicit type
conversion, when invoking '__builtin_memcpy' with a negative size
(int) value inside the library function.
The implicit type conversion moves to the function call, which still
leads to an read page fault in case of a negative size value, but
this type conversion is the fault of the user and not of the library.

Improvements:
* Range documentation for size parameter [0,L4_VCON_READ/WRITE_SIZE];
* Improves the interface, as no question about the validity of
  negative values arise.
* No need for tests with negative size values
  (they are out-of-spec anyway)
* l4_vcon_read_with_flags: one error condition check less.

The cast of the return value in `l4_vcon_write_u` from `unsigned` to
`long` is riskless, as the maximum value for size is well in range of
`long`.

Change-Id: I20c562938fe80368f07a12f66626ff6f1d290274