Ned: support automatic name-space creation
Automatically create a namespace for all tables that are referenced in
the `caps` table. This means you can write i`xxx:start({ caps = {
xyz = { aaa = ... } } }, ...)` and this would create a name-space with a
capability named `aaa` and provide this name-space as initial capability
`xyz` to the application.

This is related to ticket #495.

Change-Id: I00de454e6295a0e38e7c6fdbd3d26996643e55e6