moe: support marking boot modules as writable
This change allows the user to mark boot modules as writable.
The user can configure writable boot modules in the
modules.list by adding the parameter ':rw', like this:

module foobar :rw

Modules marked as writable are placed in a namespace called
'rwfs'. They do not show up in the 'rom' namespace, which is
reserved for read-only/copy on write dataspaces.

If mapped correctly (mind the lua script), the client can
write to the dataspace of such a boot module.

Fixes ticket #491.

Change-Id: I715fd8bde88dbbd7615f681b0243b15741412f0c