Prepopulate the page allocator with a few pages of static initial memory
In libstdc++ before v5, memory for exceptions was allocated from a
static buffer. In v5+, the static buffer was replaced in favor of
dynamically allocating a fixed-size arena during the static
initialization of emergency_pool. For moe, this is, however too early as
dynamic allocation is enabled only in find_memory() called from main().

This change allows the dynamic allocation of the memory arena to succeed
by giving the page allocator a little bit of static memory early on. In
turn, moe has an arena to allocate memory for exceptions from and can
handle out-of-memory situations gracefully.

Change-Id: I5ac6f7b7b3995bf4ebbe57df932f346a3af83d23