Revert "Make cxx::unique_ptr's of related types compatible"
This reverts commit 6a5376734348e096357ae1789524244c11ea583a.

The problem here is that the deleter used in our unique_ptr
might use the delete operator for the base class with this change,
and hence might call the wrong destructor, and even worse this might
happen implicitly.

Avoiding the automatic conversion will at least provide the burden for
the programmer to explicitly convert to the base class type and think
about the delete operator and if the correct destructor is used.

A correct solution needs to use a more complex deleter and check if
the original delete and the new deleter correctly handle the converted
data type.

Change-Id: Icb635c45a18d188fca25cd7cbd9800cf81b1fe94