Include the Kip Info memory descriptor in iomem
The info memory descriptor in Kip is used to locate the ACPI RSDP
structure by Fiasco, especially on systems that no longer support the
legacy method of discovering RSDP. In the same vein, it can be useful to
acpica, and in turn to io. For this, however, the memory range occupied
by the structure needs to reside in iomem so that acpica can map it.

Change-Id: Ib12c3d25550d9f3c1c0a974768db6812ddc0d7d3