l4sys: add function to L4::Kip::Mem_desc to retrieve all descriptors
This change adds a function to  L4::Kip::Mem_desc which returns an
enumerable list of all memory descriptors in the Kip.

This allows a couple of components such as moe, sigma0 and the
l4re_kernel to use a more concise style to iterate all memory
descriptors using a range-for loop.

Change-Id: Ic2bacfec2817f97a07601c6cb3e69c9178302206