l4sys: add `final` to generated dispatch methods
In general the dispatch methods generated by our RPC framework must not
be overriden by any user-supplied version, so we add the new `final` to
those methods if our base class uses a virtual dispatch method.
Note, we should use something like cxx::has_virtual<BASE::dispatch> or
the like but as long as this does not exist we assume if BASE is
polymorphic that dispatch is virtual in that case.

Change-Id: Id595549c6540eeb1332f98fea84a4b2f01659cf1