moe: dataspace allocation with specific range
When creating a dataspace it is now possible to specify a memory range
in which the dataspace should be allocated. This allows for explicit
allocation below 4GiB or processor-local allocation on NUMA
architectures. Allocation fails if no suitable memory block could be
found within the specified range. The default range is the full 64bit
address space.

Change-Id: I950c68a3dbfecf7f7f2d72171a6e49e7dbeec0dd