ned: Lua, make create() also return error code
If a factory create() call from Lua fails with an error returned by the
server the error gets propagated via luaL_error(). A caller may catch the
error using pcall(), however he only gets an error message and no error
code. This makes it difficult to make sense of the error e.g. whether it
is worth trying again.

This change modifies the implementation of create() such that in case of
an error it returns a table which contains an error message and error
code. Ned checks the return value of lua_dofile() whether it is a table,
extracts the error message and prints it. That retains the original
behaviour of create().

The user can now wrap the create() call using pcall() and can check for
the status. If an error happend the user can extract the message and
error code from the table and act accordingly. See the following
example:

  local status, cap = pcall(function()
                              net:create(0, 88)
                            end)
  if status == false and type(cap) == "table" then
    print(cap.msg .. " code: " .. tostring(cap.code))
    error(cap.msg)
  end

Change-Id: Ib94d3f1fc7320f4f0a4bc3e90226dccedd7337f9