uclibc/syscall: Make common implementation match unistd.h
The definition of syscall() in unistd.h is with varargs.  Traditionally
the common implementation in uclibc has been with regular arguments.
This patch updates that by using varargs.

This has caused issues on architectures like or1k which have different
calling conventions for varargs and regular arg parameters.

The implementation here is based on an implementation from Joel Stanley
<joel@jms.id.au>.  There is a difference that I do not initialize the
stack args with 0 as they are immediately overwritten by va_args.

Change-Id: I45d87bbc5eab5f08d185c1953ffe52b1013d1bbb
Signed-off-by: Stafford Horne <shorne@gmail.com>