ARM32: Use kernel's DMB function
In our libgcc we have implemented
__sync_synchronize() via 'dmb', however, we use
two variants depending on the running
core architecture. To figure that out we
need the KIP. Now it might happen that
__sync_synchronize() is called before
we could map the KIP which then breaks
our __sync_synchronize().

Anoter approach is to supply the proper
DMB variant on the syscall page, as we
do already with other calls.
This is used here.

(Arm strongly discourages to use
the mcr variant on ARMv7 and ARMv8,
thus do supply both variants.)

Change-Id: I3dabd5e24b658d41475b70089890547481bbbf60
Requires-Change-Id: If2490263dcc09c902c48cb1f93e7e482ea8dbf90