L4Re::Vfs: Use better abstraction for cap-alloc
Add the new reference counted version of the virtual capability
allocator. This allows to do correct refcounting of caps used for
mmapped files etc.

In the case of dynamic linking we first use an ldso internal allocator
and then later switch to the normal L4Re::Util allocator if the
lib4re-util is used.

This fixes ticket #213.

Change-Id: I4b5dcb08364d6225382d44a554a5f3981d6fa668
Requires-Change-Id: I6c73ea062912da6b40d87503053c57f72de675c0