pthread: fix pthread exit behaviour
Libpthread registers an exit handler using atexit(). When this is run, it
requests the pthread_manager to handle shutdown. The manager will delete all
pthreads except the caller and the main thread. We then kill the pthread
manager and the main thread (if exit() was run on a different thread).

This change implements thread killing in three steps:
- cancel ongoing IPC
- force the threads into a noop function
- delete the thread kernel object

It does not allow for graceful thread shutdown (like
pthread_cancel would do).

The previous behaviour was to call into the parent to kill the
task. Depending on the ordering of exit handlers, this caused
other handlers to not be run.

Fixes #107

Change-Id: I8d909be9162db70dd13789cdf79654aab9a6464b