test: exception handler for cache test
On ARM64 the cache operations are a user land operation. Thus a
cache-clean generated page fault is propagated to the faulting user
land thread.
On other architectures including ARM32 the cache operations are
privileged and the kernel handles the arising page fault. The introduced
exception handler is architecture agnostic.

fixes Ticket #787

Change-Id: I3ed15f8163b31f39ca53c273b7a401f4e633bb71