test/moe: create copy of data returned in UTCB
create_blocked_rm() received data in the UTCB from get_regions() and
used this data in a reserve_area() call. This call overwrites the first
Region data returned by get_regions().
The region->end value pointed to was destroyed by the reserve_area()
call but used to compute the next address to use in get_regions().
If get_regions() returned only one region, this faulty end data is
relevant as this data is directly used to query get_regions(). By now
the data at the memory location is the same as the former
region->start, leading to a loop with get_regions()-reserve_area()
calls.

fixes ticket #816

Change-Id: I4a5679b90b7ec034e3115f454bb4ca357a616244