cxx: fix license in ipc_timeout_queue
Using the LGPL-2.1 for this file is not intended as this file is
included, compiled and statically linked into the resulting binary. It
is not possible for a user to replace this later with a newer or his own
version.

Instead we put this file under the GPL-2.0-only and add the GCC runtime
library exception 2.0 to avoid implying restrictions coming from the use
of this file.

Change-Id: I4e16a280f5986fb3db2de8f32bba3dccf3bec203