uclibc/nptl: Clear TLS area for static binaries.
busybox init checks it has pid 1, but getpid() returns another value
when building busybox statically. This is because the corresponding
area is not cleared when allocated (it is allocated with
MAP_UNINITIALIZED, whose behavior depends on the Linux kernel's
CONFIG_MMAP_ALLOW_UNINITIALIZED).

This patch fixes the problem by explicitly clearing the memory area.

	* libpthread/nptl/sysdeps/generic/libc-tls.c (__libc_tls_setup):
	Clear tlsblock.

Change-Id: Id7ad69f5664847692f6e865e5875f8e59449d968
Signed-off-by: Mickaël Guêné <mickael.guene@st.com>
Signed-off-by: Christophe Lyon <christophe.lyon@st.com>