Fix htab_delete loop counter
dlclose can crash or stuck the system:

When calling dlclose() then system can crash or freeze, because
htab_delete() in ldso/ldso/fdpic/dl-inlines.h uses size_t for i which
is a typedef to unsigned int.  We exit the loop on negative value of i
which can never occur since i is an unsigned int. This leads to random
free of various pointers that kill the system.

	* ldso/include/inline-hashtab.h (htab_delete): Change type of 'i' to int.

Change-Id: I1b8e3ef626e7f645f0eac18dfa789d4841877481
Signed-off-by: Mickaël Guêné <mickael.guene@st.com>
Signed-off-by: Christophe Lyon <christophe.lyon@st.com>