l4sys/types.h: sanitize negative shift
Shifting negative values is not defined according to the C standard. We
can assume that negative values are implemented as two's complement and
in that case just doing the bit shift is the right thing. Thus we have
to cast the label to the corresponding unsigned type before shifting it.

The explicit cast back to the signed type is required as otherwise g++
complains about an implicit restriction (-Wnarrowing).

Fixes #980

Change-Id: I53c779b220852da9f3c685f5decc2f65a5c3e8ad