libstdc++-headers: gcc 7 / 8 fix like gcc 6
We do the same change here as we do for gcc 6. The presence of
GTHR_ACTIVE_PROXY is used to determine if __gthread_active_p should
return true or false. In the latter case, __gthread_mutex_lock() is
just a NOP.

With __GLIBC__ defined (this comes from <uclibc/features.h>) the code
expects a weak reference to pthread_key_create() which apparently isn't
available (0).

Fixes #979.

Change-Id: Ifda17b69ca597eacc2e8eecce9c2be6a84e0b53a