cxx: fix use of non-visible functions
The avl_tree implementation uses functions from the generic bst_node classes
that are declared protected. However, avl_tree is neither derived from
bst_node nor a friend of it.

This compiles in gcc but fails to compile in clang which enforces the
standard correctly. Since avl_tree_node imports the functions from bst_node
through a using directive and is a friend of avl_tree use these functions
to conform to the standard.

Fixes #1019

Change-Id: I7f6cc91476c497b72d428bb32647b5305879717e