Define main thread priority as a constant
This change defines the priority of the main thread of an L4Re
application centrally in l4re/consts.h. Currently this value
is used in the l4re_kernel and in libpthread.

The central definition is an implementation of the SPOT rule,
and shall ease discovery of priorities for newcomers.

This change was proposed by Christian Ludwig
<christian_ludwig@genua.de>.

Change-Id: I98f68fb9bfe84017b608baaebaf431fd1fca9cc7