Counting_cap_alloc: remove redundant if statement
The current implementation of Counting_cap_alloc::alloc()
checks twice in a row if `_free_hint >= _capacity`. The
removed if statement is redundant, because the expression
that it evaluates is evaluated again prior the first
iteration of the for loop that suceeds it. In both cases
where the evaluation occurs, if `_free_hint >= _capacity`
evaluates to true, the method terminates returning
L4_INVALID_CAP.

Change-Id: I45680e175329511ce40f4125d9681d9a1aac63a3