ned: add more named constants to L4 Lua lib
The mode() / rights() function in our L4 Lua library supports specifying
the rights using a string. It is also possible to use named constants
from the L4.Rights table. However not all possible rights have a named
constant.

This change adds named constants to L4.Rights for the 'd', 'n' and 'c'
right. This makes it consistent with the string version.

Change-Id: I99f0cae26d551cc8bba7285f6e20dcae27e74ded