Fix off by one bug in region map servers
The get_regions implementation in region_mapping_svr[_2] contains an off
by one bug in when it creates an internal Region object from the address
of a region. This change fixes the bug by using the constructor of
Region that just takes the address and creates a 1-byte region.

The bug was discovered by Diego.

Change-Id: I5275b968e2ff24ce37c1ebdeef27dcebd6d5727f