sigma0/moe: add factory documentation
Add a new page (sigma0.dox) that provides a general description
for sigma0 and contains a dedicated section for the Factory
implemented by sigma0.

Extends the Moe documentation (moe.dox) with a subsection per
each object that can be created by the Moe Factory. Information
previously contained in the section 'Memory Allocator, Generic
Factory' is moved to the Factory section, which now also
documents how to pass arguments to the create stream using <<
(for the objects supported by the Moe Factory).

This change also removes detailed information about the Mem_alloc
since it is just a convenience interface for the factory to create
Dataspaces.

Change-Id: I9faed090d6fe667eb8a67ccbb2d8c13e7b05e53b