Cleanup page_fault RPC
Fiasco newly supports the possibility to trigger exception handling
after unhandled userspace page fault via a reply with negative msgtag
label.

This commit switches to the new mechanism, paving thus the way for the
removal of the old one. It also removes the now unused optional result
argument of the page_fault RPC.

Change-Id: I6db4fa191b1dcf58536dcc33d04f28cfd51e056d