cxx/slab_alloc: remove redundant cycle from iteration
The initialization of the free list `f` in `add_slab()` updates twice
the `next` pointer of the first element of the list. The first update
(for i=0) is unnecessary, as it only serves to initilize `f`, but such
initialization is already done by the statement preceding the loop.

The initilization of the list remains correct for the corner case where
`objects_per_slab == 1`. In this case, the last element of the list is
also the first, and is initialized by the assignment `f->next = 0`, which
follows the iteration.

Change-Id: Ide2407f0717f84c4ad8c874af495188d4cb561fb