moe: Prevent array out-of-bounds access
When the flags provided by the user to Dataspace::map specify the
mapping to be both uncacheable and bufferable this may lead to an out of
bounds array access in the local cache_map array.

The behavior with this flag combination is undefined but we still add
an entry to the array to handle the case.

Change-Id: I249160bfd3033f8580f7bce54993a92a2c1c4014