l4re_kernel: add missing "0x" prefixes in exception string
The page fault address (PFA) and program loading flags (LdrFlgs)
of an unhandled page fault exception are printed in hex without
the 0x prefix, while the program counter of the exception state
(PC) is printed in hex with the 0x prefix. For making the
output consistent, the prefix 0x is added to PFA and LdrFlgs.

Change-Id: Ia0a5fee4223f24dfd30ea1c00392a23a051055f3