Make the emergency memory RWX
The memory initially used for the emergency buffer gets later used for
e.g. dataspace memory. If such dataspace maps executable memory, moe's
mapping for this memory must be also executable otherwise it cannot map
this memory as executable to its clients.

Unfortunately, the emergency memory buffer is located in moe's data
segment and as such is not executable. This commit makes sure the
emergency memory buffer is RWX by explicitly requesting the respective
pages as RWX from sigma0.

Change-Id: I1882a7f74ff15708e767f96a3ed4c924494cdbcb