moe: Add more elaborate search to detect cmdline
Up until now moe detected its command line args by searching for a
command line among the modules which contains "moe". This can lead to
a problem when another module's command line also contains the
substring "moe", e.g. within the path from the build host, within the
binary name or as another module's arguments.

To prevent selection of the wrong command line in these cases we add
further checks to the search. First we expect the word moe to be
followed by a space, iff moe has arguments, or a null-terminator, iff
moe does not have arguments, so that paths with directories containing
"moe" aren't detected as moe command line. Second we expect the word moe
to either be at the beginning of the command line or directly after a
directory seperator '/' so that modules with names ending in "moe"
aren't detected as command line. Last but not least we now do these
tests on every occurrence of "moe" within a cmdline so that an actual
moe command line with a "moe" substring within its path is not skipped
by its first occurrence.

Change-Id: I422737975855bdbdbb6aecdeeda94139a0bb641c