ned: configurable default caps per loader
Adds an optional field 'default_caps' to the loader class that
defines the capabilities that are handed to all applications
started by the given loader. If the field is not present then
the rom capability is forwarded (corresponding to the current
behaviour).

Two important use-cases:

1. For secure environments the rom capability may be removed for all
applications.

2. For debugging purposes the jdb capability may be handed in to
   all tasks.

Note that this change only introduces a mechanism so that
loaders can be built where the user has complete control over
the capabilities handed in during start(). It does not change
the current default policy. This needs to be discussed and
changed separately.

Change-Id: I16999e1066b0fdd075a713f8060a3bc8f13437a5