object_registry: document unregister requirement
Make it clear that for all register_*() calls an equivalent
unregister_obj() needs to be called. So far this was only stated
for the functions that allocate a capability. However, it is also
necessary for all other register functions because they bind
the thread. If not unbound, it may not be properly destroyed.

Change-Id: Ib03031f84a8a53a125a0ee7b26c523c3fef75a07