fix range of l4_cache_*() for ARM64 and MIPS
The last cache line of the memory range was sometimes not considered.
This could happen if we started in the middle of a cache line. For
instance
 * cache line size 0x40 bytes
 * flush 0x20-0x5f (0x40 bytes)

The old code will start with flushing the cache line 0x00-0x3f. After
the first turn, the loop address was set to 0x20 + 0x40 = 0x60 which
terminated the loop. The cache line 0x40-0x7f was not flushed despite
the fact that 0x40-0x5f is part of the flush range, too!

Change-Id: I3f9644e1b056c3f4e7f62e94ce0285f98e2335de