add reference to server right to IPC gate doc
The documentation for both the C and C++ IPC gate API
did not mention the L4_FPAGE_C_IPCGATE_SVR flag that
needs to be set on the capability in order to invoke
the control methods.

This commit fixes this and also makes two references
to other constants into a link in the generated
doxygen.

Change-Id: Iacdaf339506151feeb52f3166840c9b448ddf941