l4re-util: clean up caps in Br_manager
The Br_manager allocates capability slots where it can receive
capabilities from clients. These slots are not freed when the
Br_manager is destroyed. This causes problems in tests where we
frequently start and destroy threads with server loops.

It would be nicer to use Unique_caps here and have them handle
the resource management but the caps are cached in a convenient
array of buffer registers and we want to keep that. Write an
explicit destructor instead.

Also includes a fix of one test that checked for presence of
a capability without ensuring that the capability slot was
really empty at the beginning of the test.

Fixes #CD-193

Change-Id: Ife2a1456b5756491dbd89c7082101e8ad279b82c