Add C++ terminate handler with special L4 support
The original terminate handler form the C++ standard library print
std::exception error, but not L4::Base_exception and L4::Runtime_error
exception is a nice and verbose way. This change add a handler with
verbose support for our L4 specific exception classes.

To enable this in your program you need to include
<terminate_handler-l4> in your main.cc file.

Change-Id: I9a09039cb598d3d7ff447e7d7f3a8a7e9d02f375