rename L4Re::abort() to L4Re::throw_error()
The name 'abort' was probably not the ideal choice because there is a
libc function abort() with different semantics. Furthermore 'abort'
sounds like the thread or task is aborted but actually the L4Re C++
function throws an exception.

Reflect this meaning in the name, even if the name is longer now.

Change-Id: I6f14794adc367c9355f3f42b8501f2f8b910cfe5