moe: Use default value for non-essential parameter
In this context, there is no reason to use specifically RW flags. This
is merely a left-over from the times when the read flag also implied
executable flag and there was no separate X flag. In this regard, full
flags should be specified as RWX. Note however, that the flags
specified here are not used anywhere, so the best solution seems to be
going with the default value (i.e. RWX).

Change-Id: I22d838ad4ff0cd17e871db13d8098b59b41d262e