l4re_kernel: fix warning in release mode
In release mode the compiler warns that __libc_alloc_initial_tls does
not return a value even though the signature specifies a return value.
In debug mode this should always assert since it is not expected that
this function is ever called.

Return a nullptr in release mode to quelch the warning.

Change-Id: Id94d7ef8e00a469f608472b18047f4f810c82e01