moe: Throw correct error if entry already exists
If the mapper fails to add a node to the _map AVL tree it means that a
node with the same key is already in the tree, not that there was no
memory. This makes the two mappers throw the same error in this
situation.

Moreover this error state should not be reachable under normal
circumstances, so a comment is added.

Change-Id: I7906ee06bebc0832671f05592d5929aa064d6715