ned: implement task state lua callbacks
Currently it's only possible to wait for a single task to terminate.
This is not sufficient when managing many tasks at once. This commit
adds the ability to register a callback that is invoked when a task
terminates:

  task = L4.default_loader:start(...)
  task:exit_handler(function(exit_code)
    if exit_code == nil then
      print("Task was killed")
    else
      print("Task has terminated w/ code [" .. exit_code .. "]")
    end
  end)

If the task did already terminate then the callback is invoked
immediately. Usually the ned script should now use continuation style.
So instead of

  task = L4.default_loader:start(...)
  task:wait()
  do_more()

the script should use

  task = L4.default_loader:start(...)
  task:exit_handler(function()
    do_more()
  end)

though the former is still supported.

Change-Id: I86622193b255dac717a5dceab39e7681ea5ca32a