Moe: Dataspace_cont copy optimization
Introduce Dataspace::copy_address() delivering the biggest chunk which
can be copied. For continuous dataspaces this is the dataspace size
minus the dataspace offset.

Due to this optimization there is only a single memcpy() call when data
is copied from one Dataspace_cont to another. There is no need to copy
in flexpage chunks in that case.

The benefit on hardware is not big but measurable.

Change-Id: I6a4a155b1a4830d31d9b70f7c3717985b63dc181