cxx: implement bitfield assignments from other bitfield
Up to now it is not possible to directly assign bits from one bitfield
to another because the implementation uses an internal bit proxy type
for assignments.

This change implements the copy assignment for the internal helper type.
This allows for direct assignments from one bitfield to another e.g.

  b1.ctrl() = b2.features();

Change-Id: I76d082c782ddf05eb0d0a041275fc584cae3a92e