loader: use correct flags for stack local attach
So far the stack for the loader thread was locally attached to the
regions using default Region_handler flags of 0. This works as long as
only one (the top-most) stack page is used because that page is already
mapped before the loader thread is started.

However, if the loader thread requires more stack (for example for
throwing an exception), the region manager verifies the Region_handler
flags of the stack and complains that the stack region is not writable.

Also fix an 80-characters limit to satisfy the coding style.

This fix is a follow-up to I77273a5bb93c9891bca4f848c9b17db332b1b72a.

Change-Id: I6dd6f9877787735f44b01f0f35d9d38f7b7d5a32