moe: make Task_mapper::remove() public
Previously, Task_mapper offered a private implementation of remove()
that is an abstract public method of the Mapper base class. This worked,
because it was only called through a pointer of type Mapper, and thus
according to class.access.virt of the C++ standard, the call only checks
the permissions of the base class.

However, this is confusing, recommended against in various C++ style
guides, and most likely not a deliberate decision
(cf. Phys_mapper::remove(), which is public). Hence, move the method to
the public interface of the class.

Change-Id: I2b02e8ae28863a45589d86a7f15e654d3e6bee1e