[API] dataspace: add map_info method
In case of a MMU-less system, the dataspace must be mapped at the
correct address in the task because virtual and physical address must
match. This method returns the start and end address of the physically
contiguous buffer backing the dataspace.

On MMU-enabled system any page aligned address is permissible. On such
systems the method is just a stub.

Change-Id: I26a19bf63d769c24099dc6625bbeeba50b809da6