libloader: support ELF loading no-MMU systems
On such systems, all load sections need to be copied because there is
no paging possible. The only exception are read-only sections that are
already at the right location.

Most importantly, the load sections need to be placed at the physical
address that is recorded in the ELF file. This information must be
passed to the loader implementation that allocates the dataspace.
Unfortunately, all implementations needs to deal individually with this
problem.

The KIP dataspace needs to be handled special too. It is always present
and must not be mapped explicitly.

Change-Id: I29e446adea2b39fdd8b9ffeab9a9e35ccbd64fcb