uclibc: do not rely on syscall page for TLS initialization
On no-MMU systems there is no syscall page. Instead, call the kernel
Thread-API (l4_thread_arm_set_tpidruro()) directly. On armv6k and above
the syscall page code does exactly this anyway. The call is a no-op on
pre-armv6k architectures because there is no TLS reg to begin with.

Change-Id: I8576bad5493e48a74d42813ac714274aeda84e54