moe: rename _size and remove unused assignment
Rename _size to size and use this->size() where the method overloads the
variable. Since _size can be confused for the class member this also
better adheres to the coding standard.

Remove the unnecessary assignment to _size within clear() as it's a
remnant from when it returned a size result.

Change-Id: I4a388c438f09628028484a46f4e8db332e52fe74