l4sys: doc rights parameter of the RPC interface
The rights parameter is the first parameter of a server side
implementation of functions defined in the RPC interface.

This adds the function signature of the server side together with
a definition for the Rights bitfield.

Furthermore, the return types in the Calc example are changed to long
to be in accordance with the 'RPC Return Types' section.

Change-Id: I1c99fb99b182a599d9cdb94138fd5ed717c09fbd