Support utcb allocation on MMU-less systems
On systems without MMU the kernel effectively determines the address of
the utcb area because there is no address translation. The kernel
factory (and Task::add_ku_mem) return the allocated address in MR[0].
Adjust all involved calls to use the reported address.

The loader must be slightly refactored to allocate the task object
before populating the aux information.

Change-Id: I49d98186d21ff2c2dd83b3dedacce6a1f7fbbe88