l4re: map stuff eagerly on no-MMU systems
Even though lazy mapping is possible with an MPU it will soon lead to
the depletion of MPU regions due to fragmentation. As the memory is
fully allocated already on such systems there is no point in doing the
mapping lazy.

In case the caller must make sure that the dataspace is mapped lazily
on all systems the F::No_eager_map flag is added. This is is a no-op on
MMU enabled systems. Conversely the F::Eager_map flag makes no
difference on no-MMU systems. In case both flags are specified the
F::No_eager_map flag always wins.

Change-Id: I62de8907d1b32ae7acc00896c7b8109b9984d220