l4re: rewrite kip address handling
Having a global magic __L4_KIP_ADDR__ symbol for the address of the KIP
does not bode well with relocatable executables. The symbol will be
shifted too by the load address delta and will not match the actual KIP
mapping of the loader. Because systems without MMU have a global fixed
address of the KIP it is not useful to change the loader. Instead the
loader forwards the chosen KIP address through an aux vector entry.

The KIP pointer is stored in the l4sys library an provided by a new
l4_kip() call. In case uclibc is not used (think L4Linux) the magic
__L4_KIP_ADDR__ is still used as initial value of the KIP pointer.
Because L4Linux uses a potentially different ABI on some systems the
libl4sys-l4x.a library must be compiled with dedicated flags.

Also resolve the KIP address in the .rol4re_elf_aux section at compile
time. Otherwise the compiler will leave a relocation in the read-only
section which cannot be resolved at runtime and that will let the
linking fail.

On no-MMU platforms the KIP address is not hard coded anymore. Instead
pass it from sigma0 via moe to all tasks.

Change-Id: Ia3000a9ab3de16cca21412bf2d348195a86ebeaa
Requires-Change-Id: Iea8fc655a343028744138461a8ee040d579c81b5