l4util: Improve MBI memory map iterator macro
The iterator macro for the MBI memory map referenced a name that was not
part of the macro, but needed to be defined correctly by the user of the
macro. This change ensures that only names local to the macro are used
within it.

This change also moves parts of the iterator macro into their own
functions for better legibility and to enable other use cases that do
not use the iterator but require access to the memory map.

Change-Id: I6503cdc9b1cd17ddef54a7a353e5a33bedb740dd